pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.46.0",
    "nltk>=3.9.1",
    "openai>=1.63.2",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.9",
    "requests>=2.31.0",
    "streamlit>=1.42.0",
    "trafilatura>=2.0.0",
    "twilio>=9.4.1",
]
